Fixed conversion warnings (using too often static_cast though)

+++ src/context/StateMovingUnit.cc
@@ -50,8 +50,7 @@ StateMovingUnit::StateMovingUnit()
   float x = static_cast<float> (_holoUnitSprite->getTexture()->getSize().x);
   float y = static_cast<float> (_holoUnitSprite->getTexture()->getSize().y);
   using p = graphics::MapGraphicsProperties;
-  _holoUnitSprite->setScale(static_cast<float> (p::cellWidth()) / x,
-                            static_cast<float> (p::cellHeight()) / y);
+  _holoUnitSprite->setScale(p::cellWidth() / x, p::cellHeight() / y);
 
   // Fading sprite at original position
   auto unit(Status::battle()->map()->unit(_originalCoords));
